requirements.txt
discord.py==1.7.3
PyMySQL==1.0.2
python-dotenv==0.19.2
table2ascii==0.1.4
